refactor(prompt): move project rules loader
